Further changes to the ECS
