+++ LukkelEngine/src/LukkelEngine/Editor/Editor.cpp
@@ -1,33 +1,278 @@
 #include "LKpch.h"
 #include "LukkelEngine/Editor/Editor.h"
-#include "LukkelEngine/Scene/Components.h"
 #include "LukkelEngine/Physics/World.h"
 
 #include "imgui/imgui.h"
 #include "imgui/imgui_impl_glfw.h"
 #include "imgui/imgui_impl_opengl3.h"
+#include "imgui/imgui_internal.h"
 
 
 namespace LukkelEngine {
+
+	Entity Editor::m_SelectedEntity;
 
 	Editor::Editor(s_ptr<Scene> scene)
 		: m_Scene(scene)
 	{
 	}
 
-	void Editor::createWorldObject(unsigned int objectType, std::string_view objectName)
+	void Editor::createEntity(unsigned int objectType, std::string_view objectName)
 	{
 
 	}
 
-	void Editor::addObject(WorldObject& worldObject)
+	void Editor::addEntity(Entity& entity)
 	{
 
 	}
 
-	void selectEntity(Entity& entity)
+	void Editor::onImGuiRender()
 	{
+		ImGui::Begin("Editor menu");
+
+		if (m_Scene)
+		{
+			m_Scene->m_Registry.each([&](auto entityID)
+			{
+				Entity entity{ entityID, m_Scene.get() };
+				drawEntityNode(entity);
+			});
+
+			if (ImGui::IsMouseDown(0) && ImGui::IsWindowHovered())
+				m_SelectedEntity = {};
+
+			// Right click on blank space to get popup menu
+			if (ImGui::BeginPopupContextWindow(0, 1, false)) // id, mouse button, if on item (bool)
+			{
+				// if (ImGui::MenuItem("New entity"))
+				// 	m_Scene->createEntity("Empty Entity");
+
+				// else if (ImGui::MenuItem("New Cube"))
+				// 	// Spawner::createCube(*m_Scene, "Cube");
+				// 	void;
+
+				// else if (ImGui::MenuItem("New floor (ground object)"))
+				// 	void;
+				// 	// Spawner::createGround(*m_Scene, "Floor");
+
+				ImGui::EndPopup();
+			}
+		}
+
+		ImGui::Begin("Properties");
+		if (m_SelectedEntity)
+		{
+			drawComponents(m_SelectedEntity);
+		}
+		ImGui::End();
+
+		ImGui::End();
+	}
+
+	void Editor::selectEntity(Entity& entity)
+	{
+		if (m_SelectedEntity != entity)
+			m_SelectedEntity = entity;
+	}
+
+	void Editor::drawEntityNode(Entity entity)
+	{
+		auto& tag = entity.getComponent<TagComponent>().tag;
+
+		ImGuiTreeNodeFlags flags = ((m_SelectedEntity == entity) ? ImGuiTreeNodeFlags_Selected : 0) | ImGuiTreeNodeFlags_OpenOnArrow;
+		flags |= ImGuiTreeNodeFlags_SpanAvailWidth;
+		bool opened = ImGui::TreeNodeEx((void*)(uint64_t)(uint32_t)entity, flags, tag.c_str());
+		if (ImGui::IsItemClicked())
+		{
+			m_SelectedEntity = entity;
+		}
+
+		bool entityDeleted = false;
+		bool entityReset = false;
+
+		if (ImGui::BeginPopupContextItem())
+		{
+			if (ImGui::MenuItem("Delete object"))
+				entityDeleted= true;
+			else if (ImGui::MenuItem("Reset object"))
+				entityReset = true;
+
+			ImGui::EndPopup();
+		}
+
+		if (opened)
+		{
+			ImGuiTreeNodeFlags flags = ImGuiTreeNodeFlags_OpenOnArrow | ImGuiTreeNodeFlags_SpanAvailWidth;
+			bool opened = ImGui::TreeNodeEx((void*)9817239, flags, tag.c_str());
+			if (opened)
+				ImGui::TreePop();
+			ImGui::TreePop();
+		}
+
+		if (entityDeleted)
+		{
+			m_Scene->destroyEntity(entity);
+				if (m_SelectedEntity == entity)
+					m_SelectedEntity = {};
+			LKLOG_ERROR("Entity deleted (EDITOR)");
+		}
+	}
+
+	static void drawVec3Control(const std::string& label, glm::vec3& values, float resetValue = 0.0f, float columnWidth = 100.0f)
+	{
+		ImGuiIO& io = ImGui::GetIO();
+		auto boldFont = io.Fonts->Fonts[0];
+
+		ImGui::PushID(label.c_str());
+
+		ImGui::Columns(2);
+		ImGui::SetColumnWidth(0, columnWidth);
+		ImGui::Text(label.c_str());
+		ImGui::NextColumn();
+
+		ImGui::PushMultiItemsWidths(3, ImGui::CalcItemWidth());
+		ImGui::PushStyleVar(ImGuiStyleVar_ItemSpacing, ImVec2{ 0, 0 });
+
+		float lineHeight = GImGui->Font->FontSize + GImGui->Style.FramePadding.y * 2.0f;
+		ImVec2 buttonSize = { lineHeight + 3.0f, lineHeight };
+
+		ImGui::PushStyleColor(ImGuiCol_Button, ImVec4{ 0.8f, 0.1f, 0.15f, 1.0f });
+		ImGui::PushStyleColor(ImGuiCol_ButtonHovered, ImVec4{ 0.9f, 0.2f, 0.2f, 1.0f });
+		ImGui::PushStyleColor(ImGuiCol_ButtonActive, ImVec4{ 0.8f, 0.1f, 0.15f, 1.0f });
+		ImGui::PushFont(boldFont);
+		if (ImGui::Button("X", buttonSize))
+			values.x = resetValue;
+		ImGui::PopFont();
+		ImGui::PopStyleColor(3);
+
+		ImGui::SameLine();
+		ImGui::DragFloat("##X", &values.x, 0.1f, 0.0f, 0.0f, "%.2f");
+		ImGui::PopItemWidth();
+		ImGui::SameLine();
+
+		ImGui::PushStyleColor(ImGuiCol_Button, ImVec4{ 0.2f, 0.7f, 0.2f, 1.0f });
+		ImGui::PushStyleColor(ImGuiCol_ButtonHovered, ImVec4{ 0.3f, 0.8f, 0.3f, 1.0f });
+		ImGui::PushStyleColor(ImGuiCol_ButtonActive, ImVec4{ 0.2f, 0.7f, 0.2f, 1.0f });
+		ImGui::PushFont(boldFont);
+		if (ImGui::Button("Y", buttonSize))
+			values.y = resetValue;
+		ImGui::PopFont();
+		ImGui::PopStyleColor(3);
+
+		ImGui::SameLine();
+		ImGui::DragFloat("##Y", &values.y, 0.1f, 0.0f, 0.0f, "%.2f");
+		ImGui::PopItemWidth();
+		ImGui::SameLine();
+
+		ImGui::PushStyleColor(ImGuiCol_Button, ImVec4{ 0.1f, 0.25f, 0.8f, 1.0f });
+		ImGui::PushStyleColor(ImGuiCol_ButtonHovered, ImVec4{ 0.2f, 0.35f, 0.9f, 1.0f });
+		ImGui::PushStyleColor(ImGuiCol_ButtonActive, ImVec4{ 0.1f, 0.25f, 0.8f, 1.0f });
+		ImGui::PushFont(boldFont);
+		if (ImGui::Button("Z", buttonSize))
+			values.z = resetValue;
+		ImGui::PopFont();
+		ImGui::PopStyleColor(3);
+
+		ImGui::SameLine();
+		ImGui::DragFloat("##Z", &values.z, 0.1f, 0.0f, 0.0f, "%.2f");
+		ImGui::PopItemWidth();
+
+		ImGui::PopStyleVar();
+		ImGui::Columns(1);
+		ImGui::PopID();
+	}
+
+	template<typename T, typename UIFunction>
+	static void drawComponent(const std::string& name, Entity entity, UIFunction uiFunction)
+	{
+		const ImGuiTreeNodeFlags treeNodeFlags = ImGuiTreeNodeFlags_DefaultOpen | ImGuiTreeNodeFlags_Framed | ImGuiTreeNodeFlags_SpanAvailWidth | ImGuiTreeNodeFlags_AllowItemOverlap | ImGuiTreeNodeFlags_FramePadding;
+		if (entity.hasComponent<T>())
+		{
+			auto& component = entity.getComponent<T>();
+			ImVec2 contentRegionAvailable = ImGui::GetContentRegionAvail();
+
+			ImGui::PushStyleVar(ImGuiStyleVar_FramePadding, ImVec2{ 4, 4 });
+			float lineHeight = GImGui->Font->FontSize + GImGui->Style.FramePadding.y * 2.0f;
+			ImGui::Separator();
+			bool open = ImGui::TreeNodeEx((void*)typeid(T).hash_code(), treeNodeFlags, name.c_str());
+			ImGui::PopStyleVar(
+			);
+			ImGui::SameLine(contentRegionAvailable.x - lineHeight * 0.5f);
+			if (ImGui::Button("+", ImVec2{ lineHeight, lineHeight }))
+			{
+				ImGui::OpenPopup("ComponentSettings");
+			}
+
+			bool removeComponent = false;
+			if (ImGui::BeginPopup("ComponentSettings"))
+			{
+				if (ImGui::MenuItem("Remove component"))
+					removeComponent = true;
+
+				ImGui::EndPopup();
+			}
+
+			if (open)
+			{
+				uiFunction(component);
+				ImGui::TreePop();
+			}
+
+			if (removeComponent)
+				entity.removeComponent<T>();
+		}
+	}
+
+	template<typename T>
+	void Editor::displayAddComponentEntry(const std::string& entryName)
+	{
+		if (!m_SelectedEntity.hasComponent<T>())
+		{
+			if (ImGui::MenuItem(entryName.c_str()))
+			{
+				m_SelectedEntity.addComponent<T>();
+				ImGui::CloseCurrentPopup();
+			}
+		}
+	}
+
+	void Editor::drawComponents(Entity entity)
+	{
+		if (entity.hasComponent<TagComponent>())
+		{
+			auto& tag = entity.getComponent<TagComponent>().tag;
+
+			char buffer[256];
+			memset(buffer, 0, sizeof(buffer));
+			std::strncpy(buffer, tag.c_str(), sizeof(buffer));
+			if (ImGui::InputText("##Tag", buffer, sizeof(buffer)))
+			{
+				tag = std::string(buffer);
+			}
+		}
+		ImGui::SameLine();
+		ImGui::PushItemWidth(-1);
+
+		if (ImGui::Button("Add Component"))
+			ImGui::OpenPopup("AddComponent");
+
+		if (ImGui::BeginPopup("AddComponent"))
+		{
+			// displayAddComponentEntry<RigidBody>("Rigidbody 3D");
+			// displayAddComponentEntry<MeshComponent>("Mesh Component");
+			ImGui::EndPopup();
+		}
+
+		ImGui::PopItemWidth();
+
+		drawComponent<Mesh>("Mesh", entity, [](auto& component)
+		{
+
+		});
 
 	}
+
+
 
 }
