+++ LukkelEngine/src/LukkelEngine/Editor/Editor.h
@@ -1,8 +1,9 @@
 #pragma once
 #include "LukkelEngine/Scene/Scene.h"
 #include "LukkelEngine/Scene/Entity.h"
-#include "LukkelEngine/Model/WorldObject.h"
+#include "LukkelEngine/Scene/Components.h"
 
+#include "imgui/imgui_internal.h"
 #include "entt/entt.hpp"
 
 namespace LukkelEngine {
@@ -18,18 +19,20 @@ namespace LukkelEngine {
 		Editor(s_ptr<Scene> scene);
 		~Editor() = default;
 
-		void onImGuiRender();
+		void createEntity(unsigned int entityType, std::string_view objectName = "");
+		void addEntity(Entity& entity);
 
+		void onImGuiRender();
 		void drawEntityNode(Entity entity);
 		void drawComponents(Entity entity);
-
-		void createWorldObject(unsigned int objectType, std::string_view objectName = "");
-		void addObject(WorldObject& worldObject);
 		void selectEntity(Entity& entity);
 
+		template<typename T>
+		void displayAddComponentEntry(const std::string& entryName);
 
 	private:
 		s_ptr<Scene> m_Scene = nullptr;
+		static Entity m_SelectedEntity;
 
 		friend class World;
 	};
